Lab first commit. TODO: Add more test cases

+++ src/trianglecalculator/TriangleCalculator.java
@@ -17,16 +17,16 @@ public class TriangleCalculator {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-        int x, y, z;
+        double x, y, z;
         Scanner scanner = new Scanner(System.in);
         
         while (true) {
             System.out.println("Insert input 1:");
-            x = scanner.nextInt();
+            x = scanner.nextDouble();
             System.out.println("Insert input 2:");
-            y = scanner.nextInt();
+            y = scanner.nextDouble();
             System.out.println("Insert input 3:");
-            z = scanner.nextInt();
+            z = scanner.nextDouble();
             
             Triangle triangle = new Triangle(x,y,z);
             System.out.println("The result is a " + triangle.toString() + " triangle");
